backend/foodgram/requirements.txt
asgiref==3.6.0
Django==3.2.19
django-filter==23.2
djangorestframework==3.14.0
Pillow==9.5.0
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.5.0
